chore(code organization): move source out of index file
